refactor: Update package name and author in package.json
The code changes update the package name and author in the package.json file. The package name is changed from "@infectoone/vue-ganttastic" to "@byndcloud/vue-ganttastic" to reflect the new organization. The author field is updated to "Beyond Co. (Initial version by Marko Zunic (@zunnzunn))" to provide accurate attribution. This ensures that the package information is up to date and aligned with the current development team.